Squelch warnings.
xdtoi turns an (unsigned) 8-bit character from a string into a hex digit
value, which is also an unsigned 8-bit byte value; make it take a u_char
as an argument and return a u_char, and make the octet value accumulated
in pcap_next_etherent() a u_char as well.

Use a size_t variable for a size.