Cast away warnings.
On UN*X, a SOCKET is just an FD, hence an int, so this isn't an issue.

On Windows, well, hopefully -1, -2, and -3 aren't valid values; if they
are, this API can't be made to work (it's already documented as
returning -1, -2, and -3 for various errors).

So we might as well just cast the negative values to SOCKET, given that
we're returning a SOCKET value.

Once we support pcap_create()/pcap_activate() for remote capturing, we
can provide a better API for active-mode capturing.