Don't re-read the configuration file in a signal handler.
Instead, have the SIGHUP handler set a flag indicating that the
configuration file should be re-read, and have the main loop check that
flag and, if it's set, clear the flag and call fileconf_read().

Don't set the handler in fileconf_read() - we're using sigaction() and
telling it that the handler shouldn't be reset, so there's no need to
set it, and setting it might clear the "don't reset" indication.