Turn off non-blocking mode on the socket we got from accept().
Doing WSAEventSelect() on a socket turns on non-blocking mode on that
socket.  accept(), on a non-blocking socket, returns a non-blocking
socket and, if the listen socket is non-blocking because
WSAEventSelect() was done on it, returns a non-blocking socket *with
WSAEventSelect() in effect*, so you have to remove the events and then
turn off non-blocking mode.

Thanks, Microsoft.