Add support for running rpcapd from inetd and inetd-alikes.
When run from inetd and programs that support inetd-style program
launching (xinetd, launchd, systemd, ...), we don't do a
connection-accept loop; we're handed the input side of the connection as
the standard input and the output side of the connection as the standard
output, and should dive right into the service loop.

We extract the initial connection initialization and cleanup stuff from
daemon_serviceloop() and put it into the code that calls the service
loop, in rpcapd.c; it's different in the four cases where we start a
service loop (accepted connection on UN*X, accepted connection on
Windows, run from inetd, active mode), so that cleans up
daemon_serviceloop() a bit.

Supply a sample launchd plist; we could also supply examples of inetd
and xinetd config file entries, and systemd units.