Squelch some "no declaration before definition" warnings.
We are deliberately not declaring those variables in header files,
because they were never so declared in any libpcap API, and because we
don't want people to use them.  We only keep them exported to avoid
causing breakage for binaries that happen to use them; if, as, and when
either 1) we're sure there aren't any binaries that use them or 2)
decide we don't care whether those binaries stop working, we can get rid
of them.