Squelch a signed vs. unsigned comparison warning on FreeBSD 10.
FD_SETSIZE might be defined as an unsigned, but sock_info->sock will be
an int on UN*Xes.  It's not going to be negative (that's used for errors
from calls that return FDs, and if we got an error we wouldn't have an
FD on which to select), so just cast both sides to unsigned int.