Clean up the VLAN information handling.
Define a macro to test whether we think a packet has VLAN information,
and explain why the test is what it is (VLAN TCI != 0 or
TP_STATUS_VLAN_VALID flag set in tp_status).

If TP_STATUS_VLAN_VALID isn't defined, test the 0x10 bit (as that's how
TP_STATUS_VLAN_VALID is defined in 3.0 and later), so that if we compile
this on a system whose headers don't define TP_STATUS_VLAN_VALID
(pre-3.0 kernel headers) and run it on a system that has
TP_STATUS_VLAN_VALID (3.0 and later kernel), we'll test that flag.

In the non-memory-mapped path, make sure we fill in the aux_data
structure; that should fix Coverity CID 1430309.

Set the "VLAN information present flag" to 0 or 1; that's what the BPF
code that's generated to test the flag expects.

Set the aux_data structure members in order - "VLAN information present"
flag first, VLAN tag next.

Clean up indentation.

Add some more comments.