Add more interface flags to pcap_findalldevs().
We add:

PCAP_IF_WIRELESS, which indicates whether the interface is "wireless" or
not.

PCAP_IF_CONNECTION_STATUS, which is a bitmask for a two-bit field that
can have one of the values:

PCAP_IF_CONNECTION_STATUS_UNKNOWN if the status of whether the interface
is "connected" or "disconnected" is unknown;

PCAP_IF_CONNECTION_STATUS_CONNECTED if the interface is "connected";

PCAP_IF_CONNECTION_STATUS_DISCONNECTED if the interface is
"disconnected";

PCAP_IF_CONNECTION_STATUS_NOT_APPLICABLE if the notion of "connected" or
"disconnected" doesn't apply to this interface.

Take that into account when sorting interfaces in the interface list,
penalizing "disconnected" interfaces, as you won't see traffic on them
if they're not wireless and you'd have to be in some form of "monitor
mode" to see traffic on them if they're wireless.

This should address GitHub issue #700.