TLS for rpcap: also encrypt the control socket
This patch also encode the control sockets in adition to the data
socket. Clients performs a TLS handshake when the scheme is rpcaps://
rather than rpcap://. Both active and passive modes are supported, but
transfert via UDP is not (yet) supported (the lib returns an error in
that case).

I did some adaptation to the windows code but couldn't tested so for all
I know it may not even compile.

Also tried to fix the indentation.