Work around an annoying compiler+optimized-byte-swapping issue.
-Wunreachable-code can cause problems with ntohs(), ntohl(), htons(),
and htonl() on some platforms, such as OpenBSD 6.3 with Clang 5.0.1.  I
guess the problem is that the macro that ultimately does the
byte-swapping involves a conditional expression that tests whether the
value being swapped is a compile-time constant or not, using
__builtin_constant_p(), and, depending on whether it is, does a
compile-time swap or a run-time swap; perhaps the compiler always
considers one of the two results of the conditional expressin is never
evaluated, because the conditional check is done at compile time, and
thus always says "that expression is never executed".