Check the time stamp resolution up front before calculating with it.
ISO C says that if a left shift is done that's the full bit width of the
operand being shifted, the behavior is undefined.  In practice, it
probably just does a single shift instruction, which does whatever it
does with such a shift on the processor on which you're running; that is
likely either to be "all the bits are cleared" or "no shift is done",
but we can't rely on "all the bits are cleared".

So, instead, check the time stamp resolution exponent before using it,
and reject it if it'll result in a resolution that doesn't fit into a
32-bit quantity.

Use uint8_t for an 8-bit unsigned integral quantity while we're at it.

(cherry picked from commit a6417f7e5f259c010937b4f53490f2c552acca89)