Refine references in some man pages, round 1.
When a man page text for the first time refers to a libpcap function that
is not a topic of the man page, make sure it points to the 3PCAP section.
This way the reference becomes a hyperlink in the HTML version and the
web-site cohesion improves.

Make sure the SEE ALSO section does not list references already present
in the main text. This way the references that appear only in SEE ALSO
are much easier to notice, and cohesion remains the same.

[skip ci]