Force FormatMessageA().
Internally, we're using encodings in which ASCII characters are
represented in a single byte, so message strings must be "ANSI", not
"Unicode".  Yes, that means they'll be in the current code page, which
isn't necessarily - and probably isn't - UTF-8, but Windows really
doesn't seem to provide much support for UTF-8 as a representation of
Unicode (the current code page isn't necessarily 65001, i.e. UTF-8,
especially given that some software appears to have problems with that
code page on at least some versions of Windows).

See GitHub issue pynetwork/pypcap#64.