Set always 'time zone offset' field to zero in the pcap file header
As documented in https://www.tcpdump.org/manpages/pcap-savefile.5.txt,
this field must be 0.

It was previously propagated from the input file to the output file
when reading/writing with: tcpdump -r input.pcap -w output.pcap

Add/update the appropriate comments.

Moreover:
Set the initializations in the order of structure fields.