Add PRIsize, to use as the print format for size_t, and use it.
size_t may be bigger than unsigned long, so don't just cast to unsigned
long and print with %lu".  (In practice, the value is unlikely to be
bigger, but this is cleaner.)

We could use %zu on UN*X, as we require C99 support and C99 requires
support for z as a size modifier, but MSVC doesn't support z as a size
modifier until Visual Studio 2017.