Don't put bpf_filter_with_aux_data() in a public header file.
Don't put struct bpf_aux_data there, either.

bpf_filter_with_aux_data() is an internal-only routine, and struct
bpf_aux_data is a data structure subject to change in the future; don't
make them public, so that we *can* change it in the future without
having to worry about breaking APIs or ABIs.

(Note to any upstream packager of libpcap: if you've made it public,
so that it's exported by a shared libary, that was a mistake, so undo
it.)

(cherry picked from commit 225af6c3e8ca2fc63f20761db87af801b83d26da)