Make sure the *first* SHB, which we read when opening, isn't too big.
Start out with the 16 MiB maximum block size, check against that when
trying to read the SHB at the beginning of the file, and then increase
the maximum block size if we see an IDB with a snaplen that would result
in a bigger maximum block size.

When reading a block, we don't need to check against the maximum block
size unless we need to increase the block size.