On macOS, don't build programs any fatter than they need to be.
rpcapd need only be fat enough to run on all the platforms the target
OS version supports - no need, for example, to build it for PPC on Snow
Leopard or later, or to build it for 32-bit x86 on Lion or later - and
the test programs don't need to be fat at all.

This should fix GitHub issue #760.