Fix stats routine for text monitor.
The field giving the number of dropped URBs is "text_lost", not
"nreaders".

When scanning the values of fields, we need to get the count of
characters, just as we do when scanning the names of fields, so we know
how much to skip past the field.

Initialize text_dropped to 0 before scanning the file, so it's always
set.

(cherry picked from commit f5913533f93cb51a64771fbc2426d67f2fe571c7)