Check only VID in VLAN test (issue #461)
The SKF_AD_VLAN_TAG value in metadata contains the whole TCI so that if we
want to use it for "vlan <id>" test, we need to apply 0xfff mask to compare
only VID.

Rather than adding another 'and' instruction, use existing one we already
have for the "not offloaded" case so that when patching block like e.g.

  (010) ldh      [14]
  (011) and      #0xfff
  (012) jeq      #0xb             jt 13   jf 14

we jump to (011) instead of (012).

Fixes: 04660eb1e561 ("Use BPF extensions in compiled filters")