Redo the code that reads the configuration file.
Do a more careful parsing of the lines.  Require that they consist of:

	optional leading whitespace;

	an alphanumeric-plus-hyphen-or-underscore token, giving the
	parameter to be set;

	optional whitespace;

	an = sign;

	optional whitespace;

	a value.

If the line consists solely of leading whitespace, ignore it.

If the first character after the leading whitespace is a #, ignore the
line, treating it as a comment.

Ignore any white space after the value, *and* stop parsing the value if
you see a #, so you can put a comment after the value.

Report various syntax errors.

Use #defines for the parameter names, and use those #defines when
parsing and when saving.

This addresses GitHub issue #767.

(cherry picked from commit 5efbce7a0fe71e5d40b01a902f6febbf49c8629b)