Fix the semantics of BPF_LSH and BPF_RSH for shifts >= 32 bits.
Some processors treat shifts greater than the width of the shifted
operand as setting the destination to 0, some others treat it as a shift
modulo the width.

C says it's undefined, and most if not all implementations make it work
the way the target processor works.

We treat it as setting the destination to 0, regardless of how the
processor on which we're running works.

Credit to OSS-Fuzz for finding this issue.