Only initialize np to NULL for Linux's getnetbyname_r().
That's the only case where it's set via a pointer rather than being set
to the return address of the function, so it's the only case where it
isn't guaranteed to be set (the man page nonwithstanding).

Put in a comment explaining *why* we're setting it.

(cherry picked from commit 2954b59a01a2e5f5a3305e900659e9f2807cc875)