Add an ENABLE_SANITIZERS list variable to enable various sanitizers.
If empty, no sanitizers are enabled; otherwise, it's a
(semicolon-separated) list of sanitizers, e.g. "address;undefined", and
it will attempt to enable all the sanitizers in the list.  If any aren't
supported by the compiler, we fail.

Fix another issue we found along the way - to print a fatal error, do

    message(FATAL_ERROR "message")

not

    message(ERROR "message")