We can't longjmp out of sdup(), as it's called in the lexical analyzer.
Instead, have sdup() just set the error message to "out of memory" and
return NULL if we run out of memory.  The lexical analyzer will just set
yylval.s to NULL; the parser then has to check for all sdup()ed tokens
and, if null, YYABORT the parse.

Credit to OSS-Fuzz for finding this issue.