Fix a couple of issues.
Handle the block total length not being a multiple of 4.  Pad it to a
multiple of 4 if possible; fail if that would exceed 2^32-1.

Do the same for the total length in the trailer, and make sure the
(padded, if necessary) lengths in the header and trailer match.

(cherry picked from commit d00202a0f845129429a0f77e7dc51c6b28b2deda)