Require block lengths to be a multiple of 4.
I'm not sure what files, if any, required the Wireshark library to round
up the block length, but it was only for packet blocks, not all blocks.

We'll add it *if* there are files with which there are problems.

(cherry picked from commit cf7c3f5b17a33cff1e9ffd9b17377f857d017c7f)