Clean up signed vs. unsigned, do more error checking in the parser.
Numbers in filter expressions are unsigned; use bpf_u_int32 for them.
Use u_int for offsets and sizes that don't come from numbers in the
filter.

Have the scanner routine that parses numbers check for overflow and
report an error.

Make some routines not used outside gencode.c static.

Expand some error messages to include more details.

For 802.11 type and subtype tests with a numeric argument, make sure
we're not testing bits outside the type and subtype fields.

Credit to OSS-Fuzz for finding an integer overflow issue that the error
checking, and use of unsigned values, addresses.