Don't tell the client if we can't turn non-blocking mode off.
If the control socket is in non-blocking mode, there's no guarantee that
we can easily send an error message to the client - we *probably* can,
but let's not rely on it, especially given that there are "shouldn't
happen" errors.  Log them instead.

While we're at it, fix a copy-and-pasteo and fix indentation.

(cherry picked from commit 40b4666b567b234a5d8c2ae802c290d49758834a)