Do as much text output with rpcapd_log() as necessary.
Add a new LOGPRIO_DEBUG priority for logging, and use that instead of
SOCK_DEBUG_MESSAGE().

Use rpcapd_log() for command-line-argument errors as well.

This makes it easier to send errors somewhere other than the standard
output in daemons, and to control, at run time, whether to log debugging
messages.