Support logging rpcapd messages to the "system log".
That's what you want for daemons, as the standard error might not go
anywhere or, worse, might go to the control socket.

Also have a -D flag to control whether to log debug messages or not; the
default is "not".

(cherry picked from commit db1104b9632642bfadfa4e2e2b772ac1265fb280)