Make the key file and certificate file names local to sslutils.c.
Have routines that set them, given a pointer to the name.  Use that in
rpcapd, rather than copying to a buffer (you don't need to copy strings
from argv - unless you're going to overwrite them, which you probably
shouldn't do).

This removes a requirement for the platform to define PATH_MAX.