Pass a copy of the host/port list to daemon_serviceloop() on Windows.
Rename the sock_copy structure to params_copy, and add a pointer to the
host list to it.  Make the copy, stuff a pointer to it into the
allocated structure, and do the appropriate freeing on errors; in the
initial routine of the service thread, pass on the host list pointer
from the structure.