Don't use two sockets for the control connection.
If we do the accept() ourselves, we get only one socket on which we send
and from which we receive messages.

If we're run by an inetd-compatible daemon, it does the accept() and
gets only one socket, which it proceeds to dup and hand to us as the
standard input, output, and error; we really only need to use one of
them.

In the latter case, just dup the standard input, and then close the
standard input, output, and error as we dup a descriptor for /dev/null
to them.

In both cases, just hand the one control socket to daemon_serviceloop().
Close it in daemon_serviceloop() before it returns, rather than in the
caller after it returns.

Only free the SSL structure for the control connection right before we
close the socket for the control connection; we don't need to free it
when we close a data connection.