Get the full paths of D-Bus libraries.
At least on my DragonFly BSD virtual machine, the D-Bus libraries are
present, and pkg_check_modules() finds them, but DBUS_LIBRARIES is just
a list of library names, and, for whatever reason, the linker doesn't
look for them in /usr/local/lib.

Do the usual "we have the directories in which to look for D-Bus
libraries, and the libraries to look for; do so ourselves, and add the
full paths" stuff.