Search for "OpenSSL", not "OPENSSL".
The find module for OpenSSL packaged with CMake is FindOpenSSL.cmake.
That happens not to be a problem on my Mac, given that it's running with
a case-insensitive file system, but it fails on DragonFly BSD, which is
running with a case-sensitive file system.

(It still fails to *build* on DragonFly BSD, but that's because their
package is broken.)