Clean up the termination of a capture session.
On UN*X, instead of just cancelling the thread doing the capture, send
it a SIGUSR1, to break out of a blocking call in libpcap, and wait for
the thread to end with pthread_join().  Don't make it a detached thread,
so that we can do this.  In that thread, block SIGUSR1 unless we're in a
pcap call, so that other system calls aren't disturbed by the SIGUSR1.
(We were already doing things similarly on Windows.)  Do all session
cleanup in the main thread, after the thread completes.  That way, we
don't run the risk of trying to do cleanup in both threads.

Put the thread information into the session structure, rather than
having a separate structure.

In session_close(), if there's a thread, tell it to finish and wait for
it to finish, before cleaning up the session.  Use session_close() in
all cases where we want to finish with a capture session.

(cherry picked from commit e6715e9568197af6c1f8ce010c2f4682581db621)