Try to catch TLS to a non-TLS server or non-TLS to a TLS-only server.
That way, the failure modes for those are a bit less ugly; for the
former, you'll get a TLS handshake failure error, and for the latter,
you'll get a "server requires TLS" authentication error, rather than
some ugly protocol error from rpcapd in the former case and TLS in the
latter case.

To make this work, we just have to make sure that we never use protocol
version 22.  Do so.