More error handling cleanups.
There are several reason to avoid using gai_strerror(); write our own
routine.  Use it for all getaddrinfo() errors.

In sock_check_hostlist(), if we fail to find the host name in the list,
and we got at least one getaddrinfo() error, supply the error from that
rather than a generic "the host isn't in the host list" error.

While we're at it, just pass NULL as the service name if we're just
looking up the host name; no need to pass the string "0".

No need to report a debugging message for calls to read zero bytes;
either that's OK, in which case we should return 0, or it's not, in
which case we should stuff an error message into the error buffer and
return -1.

Don't bother reporting that we're reading and discarding data.

(cherry picked from commit 85fa933935f85c9ef462380c80833d63a106eaa2)