Get rid of the last SOCK_DEBUG_MESSAGE() calls.
Treat a failure to get the socket buffer size as an error; we were just
blithely driving on and using the variable into which the size *hadn't*
been put.  That means that our caller will be handed the error message;
we presume they do something with it, so that we don't need to print it
as a debugging message.

In other cases, we already were reporting an error and providing the
error message; the same applies there.